backend/requirement.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2
werkzeug==3.0.1